photos reorganization refactor
