--- frontend/__tests__/interface_test.ts
@@ -1,38 +1,0 @@
-import "../account/interfaces";
-import "../api/interfaces";
-import "../auth/interfaces";
-import "../config/interfaces";
-import "../connectivity/interfaces";
-import "../controls/interfaces";
-import "../controls/peripherals/interfaces";
-import "../controls/sensors/interfaces";
-import "../controls/webcam/interfaces";
-import "../devices/components/interfaces";
-import "../devices/components/fbos_settings/interfaces";
-import "../devices/interfaces";
-import "../draggable/interfaces";
-import "../farm_designer/farm_events/calendar/interfaces";
-import "../farm_designer/interfaces";
-import "../farm_designer/map/interfaces";
-import "../farmware/camera_calibration/interfaces";
-import "../farmware/camera_calibration/interfaces";
-import "../farmware/images/interfaces";
-import "../farmware/interfaces";
-import "../farmware/weed_detector/interfaces";
-import "../farmware/weed_detector/remote_env/interfaces";
-import "../front_page/interfaces";
-import "../interfaces";
-import "../logs/interfaces";
-import "../nav/interfaces";
-import "../redux/interfaces";
-import "../regimens/bulk_scheduler/interfaces";
-import "../regimens/editor/interfaces";
-import "../regimens/interfaces";
-import "../resources/interfaces";
-import "../sequences/interfaces";
-
-describe("interfaces", () => {
-  it("cant explain why coverage is 0 for interface files", () => {
-    expect(2 + 2).toBe(4);
-  });
-});
