+++ frontend/farm_designer/__tests__/photos_test.tsx
@@ -11,6 +11,8 @@ import {
   DesignerPhotosProps,
   mapStateToProps,
   ClearFarmwareData,
+  getImageJobs,
+  getCurrentImage,
 } from "../photos";
 import { fakeTimeSettings } from "../../__test_support__/fake_time_settings";
 import { fakeState } from "../../__test_support__/fake_state";
@@ -18,8 +20,9 @@ import { ExpandableHeader } from "../../ui";
 import { destroyAll } from "../../api/crud";
 import { success, error } from "../../toast/toast";
 import { fakeFarmwareManifestV1 } from "../../__test_support__/fake_farmwares";
-import { fakeWebAppConfig } from "../../__test_support__/fake_state/resources";
+import { fakeWebAppConfig, fakeImage } from "../../__test_support__/fake_state/resources";
 import { buildResourceIndex } from "../../__test_support__/resource_index_builder";
+import { JobProgress } from "farmbot";
 
 describe("<DesignerPhotos />", () => {
   const fakeProps = (): DesignerPhotosProps => ({
@@ -67,6 +70,66 @@ describe("<DesignerPhotos />", () => {
     expect(wrapper.state().detection).toEqual(true);
     headers.at(2).simulate("click");
     expect(wrapper.state().manage).toEqual(true);
+  });
+});
+
+describe("getImageJobs()", () => {
+  it("returns image upload job list", () => {
+    const allJobs = {
+      "img1.png": {
+        status: "working",
+        percent: 20,
+        unit: "percent",
+        time: "2018-11-15 18:13:21.167440Z",
+      } as JobProgress,
+      "FBOS_OTA": {
+        status: "working",
+        percent: 10,
+        unit: "percent",
+        time: "2018-11-15 17:13:21.167440Z",
+      } as JobProgress,
+      "img2.png": {
+        status: "working",
+        percent: 10,
+        unit: "percent",
+        time: "2018-11-15 19:13:21.167440Z",
+      } as JobProgress,
+    };
+    const imageJobs = getImageJobs(allJobs);
+    expect(imageJobs).toEqual([
+      {
+        status: "working",
+        percent: 10,
+        unit: "percent",
+        time: "2018-11-15 19:13:21.167440Z"
+      },
+      {
+        status: "working",
+        percent: 20,
+        unit: "percent",
+        time: "2018-11-15 18:13:21.167440Z"
+      }]);
+  });
+
+  it("handles undefined jobs", () => {
+    // tslint:disable-next-line:no-any
+    const jobs = undefined as any;
+    const imageJobs = getImageJobs(jobs);
+    expect(imageJobs).toEqual([]);
+  });
+});
+
+describe("getCurrentImage()", () => {
+  it("currentImage undefined", () => {
+    const images = [fakeImage()];
+    const currentImage = getCurrentImage(images, undefined);
+    expect(currentImage).toEqual(images[0]);
+  });
+
+  it("currentImage defined", () => {
+    const images = [fakeImage(), fakeImage()];
+    const currentImage = getCurrentImage(images, images[1].uuid);
+    expect(currentImage).toEqual(images[1]);
   });
 });
