+++ frontend/farm_designer/map/layers/images/image_layer.tsx
@@ -9,10 +9,10 @@ import { BooleanSetting } from "../../../../session_keys";
 import { GetWebAppConfigValue } from "../../../../config_storage/actions";
 import {
   parseFilterSetting, IMAGE_LAYER_CONFIG_KEYS,
-} from "../../../../farmware/images/photo_filter_settings";
+} from "../../../../photos/images/photo_filter_settings";
 import {
   imageInRange, imageIsHidden,
-} from "../../../../farmware/images/shown_in_map";
+} from "../../../../photos/images/shown_in_map";
 
 export interface ImageLayerProps {
   visible: boolean;
