+++ frontend/farm_designer/photos.tsx
@@ -2,27 +2,26 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { DesignerPanel, DesignerPanelContent } from "./designer_panel";
 import { DesignerNavTabs, Panel } from "./panel_header";
-import { UserEnv, ShouldDisplay } from "../devices/interfaces";
-import { maybeGetTimeSettings } from "../resources/selectors";
+import { UserEnv, ShouldDisplay, BotState } from "../devices/interfaces";
+import { maybeGetTimeSettings, selectAllImages } from "../resources/selectors";
 import { Everything, TimeSettings } from "../interfaces";
 import {
-  getShouldDisplayFn, getEnv, saveOrEditFarmwareEnv, getImageJobs,
-  getImages, getCurrentImage, generateFarmwareDictionary,
+  getShouldDisplayFn, getEnv, saveOrEditFarmwareEnv, generateFarmwareDictionary,
 } from "../farmware/state_to_props";
 import { JobProgress, TaggedImage, SyncStatus } from "farmbot";
 import { getStatus } from "../connectivity/reducer_support";
 import {
   prepopulateEnv, envGet,
-} from "../farmware/weed_detector/remote_env/selectors";
-import { Photos } from "../farmware/images/photos";
-import { CaptureSettings } from "../farmware/images/capture_settings";
+} from "../photos/remote_env/selectors";
+import { Photos } from "../photos/images/photos";
+import { CaptureSettings } from "../photos/images/capture_settings";
 import {
   CameraCalibration,
-} from "../farmware/camera_calibration/camera_calibration";
-import { WeedDetector } from "../farmware/weed_detector";
+} from "../photos/camera_calibration/camera_calibration";
+import { WeedDetector } from "../photos/weed_detector";
 import {
   WDENVKey, WD_ENV,
-} from "../farmware/weed_detector/remote_env/interfaces";
+} from "../photos/remote_env/interfaces";
 import { NetworkState } from "../connectivity/interfaces";
 import { t } from "../i18next_wrapper";
 import { Collapse } from "@blueprintjs/core";
@@ -36,6 +35,9 @@ import { SaveFarmwareEnv } from "../farmware/interfaces";
 import {
   getWebAppConfigValue, GetWebAppConfigValue,
 } from "../config_storage/actions";
+import { chain } from "lodash";
+import { betterCompact } from "../util";
+import { ResourceIndex } from "../resources/interfaces";
 
 export interface DesignerPhotosProps {
   dispatch: Function;
@@ -63,35 +65,56 @@ interface DesignerPhotosState {
   manage: boolean;
 }
 
+export const getImageJobs =
+  (allJobs: BotState["hardware"]["jobs"]): JobProgress[] => {
+    const jobs = allJobs || {};
+    const imageJobNames = Object.keys(jobs).filter(x => x != "FBOS_OTA");
+    const imageJobs: JobProgress[] =
+      chain(betterCompact(imageJobNames.map(x => jobs[x])))
+        .sortBy("time")
+        .reverse()
+        .value();
+    return imageJobs;
+  };
+
+const getImages = (ri: ResourceIndex): TaggedImage[] =>
+  chain(selectAllImages(ri))
+    .sortBy(x => x.body.id)
+    .reverse()
+    .value();
+
+export const getCurrentImage =
+  (images: TaggedImage[], currentImgUuid: string | undefined): TaggedImage => {
+    const firstImage = images[0];
+    const currentImage =
+      images.filter(i => i.uuid === currentImgUuid)[0] || firstImage;
+    return currentImage;
+  };
+
 export const mapStateToProps = (props: Everything): DesignerPhotosProps => {
   const images = getImages(props.resources.index);
-  const currentImageUuid = props.resources.consumers.farmware.currentImage;
-  const currentImage = getCurrentImage(images, currentImageUuid);
-
   const shouldDisplay = getShouldDisplayFn(props.resources.index, props.bot);
   const env = getEnv(props.resources.index, shouldDisplay, props.bot);
-
-  const botToMqttStatus = getStatus(props.bot.connectivity.uptime["bot.mqtt"]);
-  const syncStatus = props.bot.hardware.informational_settings.sync_status;
 
   const versions: Record<string, string> = {};
   Object.entries(generateFarmwareDictionary(props.bot, props.resources.index))
     .map(([farmwareName, manifest]) =>
       versions[farmwareName] = manifest.meta.version);
 
+  const currentImageUuid = props.resources.consumers.farmware.currentImage;
   const {
     hiddenImages, shownImages, hideUnShownImages, alwaysHighlightImage,
   } = props.resources.consumers.farm_designer;
 
   return {
     timeSettings: maybeGetTimeSettings(props.resources.index),
-    botToMqttStatus,
+    botToMqttStatus: getStatus(props.bot.connectivity.uptime["bot.mqtt"]),
     wDEnv: prepopulateEnv(env),
     env,
     dispatch: props.dispatch,
-    currentImage,
+    currentImage: getCurrentImage(images, currentImageUuid),
     images,
-    syncStatus,
+    syncStatus: props.bot.hardware.informational_settings.sync_status,
     shouldDisplay,
     saveFarmwareEnv: saveOrEditFarmwareEnv(props.resources.index),
     imageJobs: getImageJobs(props.bot.hardware.jobs),
