+++ frontend/farmware/interfaces.ts
@@ -1,11 +1,4 @@
-import {
-  Dictionary, FarmwareConfig, TaggedImage, SyncStatus,
-  TaggedFarmwareInstallation, JobProgress,
-} from "farmbot";
-import { UserEnv, ShouldDisplay } from "../devices/interfaces";
-import { NetworkState } from "../connectivity/interfaces";
-import { TimeSettings } from "../interfaces";
-import { GetWebAppConfigValue } from "../config_storage/actions";
+import { Dictionary, FarmwareConfig } from "farmbot";
 
 export interface FarmwareState {
   currentFarmware: string | undefined;
@@ -29,25 +22,6 @@ export interface FarmwareManifestInfo {
     version: string;
     author: string;
   }
-}
-
-export interface FarmwareProps {
-  dispatch: Function;
-  env: UserEnv;
-  images: TaggedImage[];
-  currentImage: TaggedImage | undefined;
-  botToMqttStatus: NetworkState;
-  farmwares: Farmwares;
-  timeSettings: TimeSettings;
-  syncStatus: SyncStatus | undefined;
-  getConfigValue: GetWebAppConfigValue;
-  firstPartyFarmwareNames: string[];
-  currentFarmware: string | undefined;
-  shouldDisplay: ShouldDisplay;
-  saveFarmwareEnv: SaveFarmwareEnv;
-  taggedFarmwareInstallations: TaggedFarmwareInstallation[];
-  imageJobs: JobProgress[];
-  infoOpen: boolean;
 }
 
 /** Function to save a Farmware env variable to the API. */
