+++ frontend/farmware/state_to_props.ts
@@ -1,30 +1,16 @@
-import { Everything } from "../interfaces";
-import {
-  selectAllImages, maybeGetDevice, maybeGetTimeSettings,
-} from "../resources/selectors";
-import {
-  Feature, UserEnv, ShouldDisplay, BotState,
-} from "../devices/interfaces";
+import { maybeGetDevice } from "../resources/selectors";
+import { Feature, UserEnv, ShouldDisplay, BotState } from "../devices/interfaces";
 import {
   selectAllFarmwareEnvs, selectAllFarmwareInstallations,
 } from "../resources/selectors_by_kind";
-import {
-  determineInstalledOsVersion,
-  createShouldDisplayFn,
-  betterCompact,
-} from "../util";
+import { determineInstalledOsVersion, createShouldDisplayFn } from "../util";
 import { ResourceIndex } from "../resources/interfaces";
-import { TaggedFarmwareEnv, JobProgress, TaggedImage } from "farmbot";
+import { TaggedFarmwareEnv } from "farmbot";
 import { save, edit, initSave } from "../api/crud";
-import { chain } from "lodash";
-import {
-  FarmwareManifestInfo, Farmwares, SaveFarmwareEnv, FarmwareProps,
-} from "./interfaces";
+import { FarmwareManifestInfo, Farmwares, SaveFarmwareEnv } from "./interfaces";
 import { manifestInfo, manifestInfoPending } from "./generate_manifest_info";
 import { t } from "../i18next_wrapper";
-import { getStatus } from "../connectivity/reducer_support";
 import { DevSettings } from "../account/dev/dev_support";
-import { getWebAppConfigValue } from "../config_storage/actions";
 
 /** Edit an existing Farmware env variable or add a new one. */
 export const saveOrEditFarmwareEnv = (ri: ResourceIndex): SaveFarmwareEnv =>
@@ -44,7 +30,7 @@ export const saveOrEditFarmwareEnv = (ri: ResourceIndex): SaveFarmwareEnv =>
 export const isPendingInstallation = (farmware: FarmwareManifestInfo | undefined) =>
   !farmware || farmware.installation_pending;
 
-export const reduceFarmwareEnv =
+const reduceFarmwareEnv =
   (ri: ResourceIndex): UserEnv => {
     const farmwareEnv: UserEnv = {};
     selectAllFarmwareEnvs(ri)
@@ -99,69 +85,3 @@ export const generateFarmwareDictionary = (
     });
   return farmwares;
 };
-
-export const getImageJobs =
-  (allJobs: BotState["hardware"]["jobs"]): JobProgress[] => {
-    const jobs = allJobs || {};
-    const imageJobNames = Object.keys(jobs).filter(x => x != "FBOS_OTA");
-    const imageJobs: JobProgress[] =
-      chain(betterCompact(imageJobNames.map(x => jobs[x])))
-        .sortBy("time")
-        .reverse()
-        .value();
-    return imageJobs;
-  };
-
-export const getImages = (ri: ResourceIndex): TaggedImage[] =>
-  chain(selectAllImages(ri))
-    .sortBy(x => x.body.id)
-    .reverse()
-    .value();
-
-export const getCurrentImage =
-  (images: TaggedImage[], currentImgUuid: string | undefined): TaggedImage => {
-    const firstImage = images[0];
-    const currentImage =
-      images.filter(i => i.uuid === currentImgUuid)[0] || firstImage;
-    return currentImage;
-  };
-
-export function mapStateToProps(props: Everything): FarmwareProps {
-  const images = getImages(props.resources.index);
-  const currentImageUuid = props.resources.consumers.farmware.currentImage;
-  const currentImage = getCurrentImage(images, currentImageUuid);
-
-  const { currentFarmware, firstPartyFarmwareNames, infoOpen } =
-    props.resources.consumers.farmware;
-
-  const shouldDisplay = getShouldDisplayFn(props.resources.index, props.bot);
-  const env = getEnv(props.resources.index, shouldDisplay, props.bot);
-
-  const taggedFarmwareInstallations =
-    selectAllFarmwareInstallations(props.resources.index);
-
-  const farmwares =
-    generateFarmwareDictionary(props.bot, props.resources.index, shouldDisplay);
-
-  const botToMqttStatus = getStatus(props.bot.connectivity.uptime["bot.mqtt"]);
-  const syncStatus = props.bot.hardware.informational_settings.sync_status;
-
-  return {
-    timeSettings: maybeGetTimeSettings(props.resources.index),
-    currentFarmware,
-    farmwares,
-    botToMqttStatus,
-    env,
-    dispatch: props.dispatch,
-    currentImage,
-    images,
-    syncStatus,
-    getConfigValue: getWebAppConfigValue(() => props),
-    firstPartyFarmwareNames,
-    shouldDisplay,
-    saveFarmwareEnv: saveOrEditFarmwareEnv(props.resources.index),
-    taggedFarmwareInstallations,
-    imageJobs: getImageJobs(props.bot.hardware.jobs),
-    infoOpen,
-  };
-}
