--- frontend/farmware/weed_detector/remote_env/__tests__/actions_test.ts
@@ -1,18 +1,0 @@
-jest.mock("../../../../device", () => {
-  const mock = { setUserEnv: jest.fn(() => Promise.resolve()) };
-  return { getDevice: () => mock };
-});
-
-import { envSave } from "../actions";
-import { getDevice } from "../../../../device";
-
-describe("actions", () => {
-  it("Saves environment variables", () => {
-    /** This test is just here to make sure that envSave() is actually
-     * triggering side effects. */
-    envSave("CAMERA_CALIBRATION_blur", 24);
-    expect(getDevice().setUserEnv).toHaveBeenCalledTimes(1);
-    expect(getDevice().setUserEnv)
-      .toHaveBeenLastCalledWith({ "CAMERA_CALIBRATION_blur": "24" });
-  });
-});
