+++ frontend/photos/camera_calibration/__tests__/actions_test.ts
