+++ frontend/photos/camera_calibration/__tests__/config_test.tsx
