+++ frontend/photos/camera_calibration/actions.ts
