+++ frontend/photos/camera_calibration/constants.ts
@@ -1,5 +1,5 @@
 import { DropDownItem } from "../../ui/index";
-import { SPECIAL_VALUES } from "../weed_detector/remote_env/constants";
+import { SPECIAL_VALUES } from "../remote_env/constants";
 import { t } from "../../i18next_wrapper";
 
 /** Mapping of SPECIAL_VALUE numeric codes into corresponding drop down items. */
