+++ frontend/photos/image_workspace/__tests__/farmbot_picker_test.tsx
