+++ frontend/photos/image_workspace/__tests__/slider_test.tsx
