+++ frontend/photos/image_workspace/farmbot_picker.tsx
