+++ frontend/photos/image_workspace/slider.tsx
