+++ frontend/photos/images/__tests__/actions_test.ts
