+++ frontend/photos/images/__tests__/capture_settings_test.tsx
