+++ frontend/photos/images/__tests__/image_filter_menu_test.tsx
