+++ frontend/photos/images/__tests__/image_flipper_test.tsx
