+++ frontend/photos/images/__tests__/photo_filter_settings_test.tsx
