+++ frontend/photos/images/__tests__/photos_test.tsx
