+++ frontend/photos/images/__tests__/shown_in_map_test.tsx
