+++ frontend/photos/images/actions.ts
