+++ frontend/photos/images/image_filter_menu.tsx
