+++ frontend/photos/images/image_flipper.tsx
