+++ frontend/photos/images/interfaces.ts
