+++ frontend/photos/images/photo_filter_settings.tsx
