+++ frontend/photos/images/shown_in_map.tsx
