+++ frontend/photos/remote_env/__tests__/actions_test.ts
@@ -1,0 +1,14 @@
+const mockDevice = { setUserEnv: jest.fn(() => Promise.resolve()) };
+jest.mock("../../../device", () => ({ getDevice: () => mockDevice }));
+
+import { envSave } from "../actions";
+import { getDevice } from "../../../device";
+
+describe("envSave()", () => {
+  it("saves environment variables", () => {
+    envSave("CAMERA_CALIBRATION_blur", 24);
+    expect(getDevice().setUserEnv).toHaveBeenCalledTimes(1);
+    expect(getDevice().setUserEnv)
+      .toHaveBeenLastCalledWith({ "CAMERA_CALIBRATION_blur": "24" });
+  });
+});
