+++ frontend/photos/remote_env/__tests__/constants_test.ts
