+++ frontend/photos/remote_env/__tests__/selectors_test.ts
