+++ frontend/photos/remote_env/__tests__/translators_test.ts
