+++ frontend/photos/remote_env/constants.ts
