+++ frontend/photos/weed_detector/__tests__/actions_test.ts
