+++ frontend/photos/weed_detector/__tests__/index_test.tsx
