+++ frontend/photos/weed_detector/actions.tsx
